+++ secure-minify/js/serverlist.js
@@ -46,6 +46,8 @@ document.getElementById('saveNewButton').removeAttribute('hidden');document.getE
 ;document.getElementById('verifyCheckboxId').checked=true;document.getElementById('proxyCheckboxId').checked=false;document.getElementById('autoReconnectCheckboxId').checked=false
 ;document.getElementById('loggingCheckboxId').checked=false;document.getElementById('passwordInputId').setAttribute('disabled','');document.getElementById('passwordInputId').value='(Blank)'
 ;document.getElementById('replacePasswordButton').removeAttribute('hidden');document.getElementById('serverPasswordWarningDiv').setAttribute('hidden','')
+;document.getElementById('saslUsernameInputId').value='';document.getElementById('saslPasswordInputId').setAttribute('disabled','');document.getElementById('saslPasswordInputId').value='(Blank)'
+;document.getElementById('replaceSaslPasswordButton').removeAttribute('hidden');document.getElementById('serverSaslPasswordWarningDiv').setAttribute('hidden','')
 ;document.getElementById('identifyNickInputId').value='';document.getElementById('identifyCommandInputId').setAttribute('disabled','');document.getElementById('identifyCommandInputId').value='(blank)'
 ;document.getElementById('replaceIdentifyCommandButton').removeAttribute('hidden');document.getElementById('nickservCommandWarningDiv').setAttribute('hidden','')
 ;document.getElementById('nickInputId').value='';document.getElementById('altNickInputId').value='';document.getElementById('recoverNickCheckboxId').checked=false
@@ -63,7 +65,10 @@ document.getElementById('saveNewButton').removeAttribute('hidden');document.getE
 ;if(data.logging)document.getElementById('loggingCheckboxId').checked=true;else document.getElementById('loggingCheckboxId').checked=false
 ;document.getElementById('passwordInputId').setAttribute('disabled','');document.getElementById('replacePasswordButton').removeAttribute('hidden')
 ;if(null===data.password)document.getElementById('passwordInputId').value='(hidden)';else document.getElementById('passwordInputId').value='(blank)'
-;document.getElementById('serverPasswordWarningDiv').setAttribute('hidden','');document.getElementById('identifyNickInputId').value=data.identifyNick
+;document.getElementById('serverPasswordWarningDiv').setAttribute('hidden','');document.getElementById('saslUsernameInputId').value=data.saslUsername
+;document.getElementById('saslPasswordInputId').setAttribute('disabled','');document.getElementById('replaceSaslPasswordButton').removeAttribute('hidden')
+;if(null===data.saslPassword)document.getElementById('saslPasswordInputId').value='(hidden)';else document.getElementById('saslPasswordInputId').value='(blank)'
+;document.getElementById('serverSaslPasswordWarningDiv').setAttribute('hidden','');document.getElementById('identifyNickInputId').value=data.identifyNick
 ;document.getElementById('identifyCommandInputId').setAttribute('disabled','');document.getElementById('replaceIdentifyCommandButton').removeAttribute('hidden')
 ;if(null===data.identifyCommand)document.getElementById('identifyCommandInputId').value='(hidden)';else document.getElementById('identifyCommandInputId').value='(blank)'
 ;document.getElementById('nickservCommandWarningDiv').setAttribute('hidden','');document.getElementById('nickInputId').value=data.nick;document.getElementById('altNickInputId').value=data.altNick
@@ -87,6 +92,8 @@ _showError(err.toString()||err);console.log(err)})};const parseFormInputValues=(
 ;if(document.getElementById('autoReconnectCheckboxId').checked)data.reconnect=true;else data.reconnect=false
 ;if(document.getElementById('loggingCheckboxId').checked)data.logging=true;else data.logging=false
 ;if(!document.getElementById('passwordInputId').hasAttribute('disabled'))data.password=document.getElementById('passwordInputId').value
+;data.saslUsername=document.getElementById('saslUsernameInputId').value
+;if(!document.getElementById('saslPasswordInputId').hasAttribute('disabled'))data.saslPassword=document.getElementById('saslPasswordInputId').value
 ;data.identifyNick=document.getElementById('identifyNickInputId').value
 ;if(!document.getElementById('identifyCommandInputId').hasAttribute('disabled'))data.identifyCommand=document.getElementById('identifyCommandInputId').value
 ;data.nick=document.getElementById('nickInputId').value;data.altNick=document.getElementById('altNickInputId').value
@@ -98,14 +105,14 @@ _showError(err.toString()||err);console.log(err)})};const parseFormInputValues=(
 ;if(''===data.user)errorStr='Unix ident user is required input.';if(''===data.real)errorStr='Real Name is required input.';if(errorStr)reject(new Error(errorStr));else resolve({data:data,index:index})
 });const buildServerListTable=data=>new Promise((resolve,reject)=>{const tableNode=document.getElementById('tbodyId');while(tableNode.firstChild)tableNode.removeChild(tableNode.firstChild)
 ;const columnTitles=[];columnTitles.push('Index');columnTitles.push('Disabled');columnTitles.push('Group');columnTitles.push('Label');columnTitles.push('Host');columnTitles.push('Port')
-;if(full)columnTitles.push('TLS');if(full)columnTitles.push('verify');if(full)columnTitles.push('proxy');if(full)columnTitles.push('password');columnTitles.push('Nick')
-;if(full)columnTitles.push('Alternate');if(full)columnTitles.push('Recover');if(full)columnTitles.push('user');if(full)columnTitles.push('Real Name');if(full)columnTitles.push('Modes')
-;if(full)columnTitles.push('Channels');if(full)columnTitles.push('identifyNick');if(full)columnTitles.push('command');if(full)columnTitles.push('reconnect');if(full)columnTitles.push('logging')
-;if(editable)columnTitles.push('');if(editable)columnTitles.push('');if(editable)columnTitles.push('');if(editable)columnTitles.push('');const titleRowEl=document.createElement('tr')
-;columnTitles.forEach(titleName=>{const tdEl=document.createElement('td');tdEl.textContent=titleName;titleRowEl.appendChild(tdEl)});tableNode.appendChild(titleRowEl)
-;if(Array.isArray(data)&&data.length>0)for(let i=0;i<data.length;i++){const rowEl=document.createElement('tr');rowEl.setAttribute('index',i.toString())
-;if(data[i].disabled)rowEl.classList.add('disabled-tr');const td01El=document.createElement('td');td01El.textContent=i.toString();rowEl.appendChild(td01El);const td10El=document.createElement('td')
-;const disabledCheckboxEl=document.createElement('input');disabledCheckboxEl.setAttribute('type','checkbox')
+;if(full)columnTitles.push('TLS');if(full)columnTitles.push('verify');if(full)columnTitles.push('proxy');if(full)columnTitles.push('password');if(full)columnTitles.push('sasl username')
+;if(full)columnTitles.push('sasl password');columnTitles.push('Nick');if(full)columnTitles.push('Alternate');if(full)columnTitles.push('Recover');if(full)columnTitles.push('user')
+;if(full)columnTitles.push('Real Name');if(full)columnTitles.push('Modes');if(full)columnTitles.push('Channels');if(full)columnTitles.push('identifyNick');if(full)columnTitles.push('command')
+;if(full)columnTitles.push('reconnect');if(full)columnTitles.push('logging');if(editable)columnTitles.push('');if(editable)columnTitles.push('');if(editable)columnTitles.push('')
+;if(editable)columnTitles.push('');const titleRowEl=document.createElement('tr');columnTitles.forEach(titleName=>{const tdEl=document.createElement('td');tdEl.textContent=titleName
+;titleRowEl.appendChild(tdEl)});tableNode.appendChild(titleRowEl);if(Array.isArray(data)&&data.length>0)for(let i=0;i<data.length;i++){const rowEl=document.createElement('tr')
+;rowEl.setAttribute('index',i.toString());if(data[i].disabled)rowEl.classList.add('disabled-tr');const td01El=document.createElement('td');td01El.textContent=i.toString();rowEl.appendChild(td01El)
+;const td10El=document.createElement('td');const disabledCheckboxEl=document.createElement('input');disabledCheckboxEl.setAttribute('type','checkbox')
 ;if(editable)disabledCheckboxEl.removeAttribute('disabled');else disabledCheckboxEl.setAttribute('disabled','');disabledCheckboxEl.checked=data[i].disabled;td10El.appendChild(disabledCheckboxEl)
 ;rowEl.appendChild(td10El);const td11El=document.createElement('td');if('group'in data[i])td11El.textContent=data[i].group;else td11El.textContent=0
 ;if('group'in data[i]&&data[i].group>0&&data[i].group<6)td11El.classList.add('group-color-'+data[i].group.toString());rowEl.appendChild(td11El);const td12El=document.createElement('td')
@@ -118,31 +125,33 @@ verifyIconEl.classList.add('icon-true');verifyIconInnerEl.classList.add('icon-in
 td23El.appendChild(verifyIconEl);rowEl.appendChild(td23El);const td24El=document.createElement('td');const proxyIconEl=document.createElement('div')
 ;const proxyIconInnerEl=document.createElement('div');proxyIconEl.appendChild(proxyIconInnerEl);if(data[i].proxy){proxyIconEl.classList.add('icon-true')
 ;proxyIconInnerEl.classList.add('icon-inner-true')}else{proxyIconEl.classList.add('icon-false');proxyIconInnerEl.classList.add('icon-inner-false')}td24El.appendChild(proxyIconEl)
-;rowEl.appendChild(td24El);const td25El=document.createElement('td');if(null===data[i].password)td25El.textContent='(hidden)';else td25El.textContent='(blank)';rowEl.appendChild(td25El)}
-const td30El=document.createElement('td');td30El.textContent=data[i].nick;rowEl.appendChild(td30El);if(full){const td31El=document.createElement('td');td31El.textContent=data[i].altNick
-;rowEl.appendChild(td31El);const td32El=document.createElement('td');const recoverNickIconEl=document.createElement('div');const recoverNickIconInnerEl=document.createElement('div')
-;recoverNickIconEl.appendChild(recoverNickIconInnerEl);if(data[i].recoverNick){recoverNickIconEl.classList.add('icon-true');recoverNickIconInnerEl.classList.add('icon-inner-true')}else{
-recoverNickIconEl.classList.add('icon-false');recoverNickIconInnerEl.classList.add('icon-inner-false')}td32El.appendChild(recoverNickIconEl);rowEl.appendChild(td32El)
-;const td33El=document.createElement('td');td33El.textContent=data[i].user;rowEl.appendChild(td33El);const td34El=document.createElement('td');td34El.textContent=data[i].real;rowEl.appendChild(td34El)
-;const td35El=document.createElement('td');td35El.textContent=data[i].modes;rowEl.appendChild(td35El);const td40El=document.createElement('td');data[i].channelList.split(',').forEach(channel=>{
-const chanDiv=document.createElement('div');chanDiv.textContent=channel;td40El.appendChild(chanDiv)});rowEl.appendChild(td40El);const td50El=document.createElement('td')
-;td50El.textContent=data[i].identifyNick;rowEl.appendChild(td50El);const td51El=document.createElement('td')
-;if(null===data[i].identifyCommand)td51El.textContent='(hidden)';else td51El.textContent='(blank)';rowEl.appendChild(td51El);const td60El=document.createElement('td')
-;const reconnectIconEl=document.createElement('div');const reconnectIconInnerEl=document.createElement('div');reconnectIconEl.appendChild(reconnectIconInnerEl);if(data[i].reconnect){
-reconnectIconEl.classList.add('icon-true');reconnectIconInnerEl.classList.add('icon-inner-true')}else{reconnectIconEl.classList.add('icon-false');reconnectIconInnerEl.classList.add('icon-inner-false')
-}td60El.appendChild(reconnectIconEl);rowEl.appendChild(td60El);const td61El=document.createElement('td');const loggingIconEl=document.createElement('div')
-;const loggingIconInnerEl=document.createElement('div');loggingIconEl.appendChild(loggingIconInnerEl);if(data[i].logging){loggingIconEl.classList.add('icon-true')
-;loggingIconInnerEl.classList.add('icon-inner-true')}else{loggingIconEl.classList.add('icon-false');loggingIconInnerEl.classList.add('icon-inner-false')}td61El.appendChild(loggingIconEl)
-;rowEl.appendChild(td61El)}if(editable){const td70El=document.createElement('td');const editButtonEl=document.createElement('button');editButtonEl.textContent='Edit';td70El.appendChild(editButtonEl)
-;rowEl.appendChild(td70El);const td71El=document.createElement('td');const copyButtonEl=document.createElement('button');copyButtonEl.textContent='Copy';td71El.appendChild(copyButtonEl)
-;rowEl.appendChild(td71El);const td72El=document.createElement('td');const deleteButtonEl=document.createElement('button');deleteButtonEl.textContent='Delete';td72El.appendChild(deleteButtonEl)
-;rowEl.appendChild(td72El);const td73El=document.createElement('td');const moveUpButtonEl=document.createElement('button');moveUpButtonEl.textContent='move-up'
-;if(i>0)td73El.appendChild(moveUpButtonEl);rowEl.appendChild(td73El);disabledCheckboxEl.addEventListener('click',()=>{toggleDisabled(parseInt(rowEl.getAttribute('index')))})
-;editButtonEl.addEventListener('click',()=>{openIrcServerEdit(parseInt(rowEl.getAttribute('index')))});copyButtonEl.addEventListener('click',()=>{
-copyIrcServerToNew(parseInt(rowEl.getAttribute('index')))});deleteButtonEl.addEventListener('click',()=>{deleteIrcServer(parseInt(rowEl.getAttribute('index')))})
-;if(i>0)moveUpButtonEl.addEventListener('click',()=>{moveUpInList(parseInt(rowEl.getAttribute('index')))})}tableNode.appendChild(rowEl)}resolve(null)})
-;const checkForApiError=data=>new Promise((resolve,reject)=>{if('success'===data.status)resolve(null);else reject(new Error('PATCH API did not return success status flag'))})
-;const setDivVisibility=data=>{document.getElementById('listVisibilityDiv').removeAttribute('hidden','');document.getElementById('formVisibilityDiv').setAttribute('hidden','')
+;rowEl.appendChild(td24El);const td25El=document.createElement('td');if(null===data[i].password)td25El.textContent='(hidden)';else td25El.textContent='(blank)';rowEl.appendChild(td25El)
+;const td26El=document.createElement('td');td26El.textContent=data[i].saslUsername;rowEl.appendChild(td26El);const td27El=document.createElement('td')
+;if(null===data[i].saslPassword)td27El.textContent='(hidden)';else td27El.textContent='(blank)';rowEl.appendChild(td27El)}const td30El=document.createElement('td');td30El.textContent=data[i].nick
+;rowEl.appendChild(td30El);if(full){const td31El=document.createElement('td');td31El.textContent=data[i].altNick;rowEl.appendChild(td31El);const td32El=document.createElement('td')
+;const recoverNickIconEl=document.createElement('div');const recoverNickIconInnerEl=document.createElement('div');recoverNickIconEl.appendChild(recoverNickIconInnerEl);if(data[i].recoverNick){
+recoverNickIconEl.classList.add('icon-true');recoverNickIconInnerEl.classList.add('icon-inner-true')}else{recoverNickIconEl.classList.add('icon-false')
+;recoverNickIconInnerEl.classList.add('icon-inner-false')}td32El.appendChild(recoverNickIconEl);rowEl.appendChild(td32El);const td33El=document.createElement('td');td33El.textContent=data[i].user
+;rowEl.appendChild(td33El);const td34El=document.createElement('td');td34El.textContent=data[i].real;rowEl.appendChild(td34El);const td35El=document.createElement('td')
+;td35El.textContent=data[i].modes;rowEl.appendChild(td35El);const td40El=document.createElement('td');data[i].channelList.split(',').forEach(channel=>{const chanDiv=document.createElement('div')
+;chanDiv.textContent=channel;td40El.appendChild(chanDiv)});rowEl.appendChild(td40El);const td50El=document.createElement('td');td50El.textContent=data[i].identifyNick;rowEl.appendChild(td50El)
+;const td51El=document.createElement('td');if(null===data[i].identifyCommand)td51El.textContent='(hidden)';else td51El.textContent='(blank)';rowEl.appendChild(td51El)
+;const td60El=document.createElement('td');const reconnectIconEl=document.createElement('div');const reconnectIconInnerEl=document.createElement('div')
+;reconnectIconEl.appendChild(reconnectIconInnerEl);if(data[i].reconnect){reconnectIconEl.classList.add('icon-true');reconnectIconInnerEl.classList.add('icon-inner-true')}else{
+reconnectIconEl.classList.add('icon-false');reconnectIconInnerEl.classList.add('icon-inner-false')}td60El.appendChild(reconnectIconEl);rowEl.appendChild(td60El)
+;const td61El=document.createElement('td');const loggingIconEl=document.createElement('div');const loggingIconInnerEl=document.createElement('div');loggingIconEl.appendChild(loggingIconInnerEl)
+;if(data[i].logging){loggingIconEl.classList.add('icon-true');loggingIconInnerEl.classList.add('icon-inner-true')}else{loggingIconEl.classList.add('icon-false')
+;loggingIconInnerEl.classList.add('icon-inner-false')}td61El.appendChild(loggingIconEl);rowEl.appendChild(td61El)}if(editable){const td70El=document.createElement('td')
+;const editButtonEl=document.createElement('button');editButtonEl.textContent='Edit';td70El.appendChild(editButtonEl);rowEl.appendChild(td70El);const td71El=document.createElement('td')
+;const copyButtonEl=document.createElement('button');copyButtonEl.textContent='Copy';td71El.appendChild(copyButtonEl);rowEl.appendChild(td71El);const td72El=document.createElement('td')
+;const deleteButtonEl=document.createElement('button');deleteButtonEl.textContent='Delete';td72El.appendChild(deleteButtonEl);rowEl.appendChild(td72El);const td73El=document.createElement('td')
+;const moveUpButtonEl=document.createElement('button');moveUpButtonEl.textContent='move-up';if(i>0)td73El.appendChild(moveUpButtonEl);rowEl.appendChild(td73El)
+;disabledCheckboxEl.addEventListener('click',()=>{toggleDisabled(parseInt(rowEl.getAttribute('index')))});editButtonEl.addEventListener('click',()=>{
+openIrcServerEdit(parseInt(rowEl.getAttribute('index')))});copyButtonEl.addEventListener('click',()=>{copyIrcServerToNew(parseInt(rowEl.getAttribute('index')))})
+;deleteButtonEl.addEventListener('click',()=>{deleteIrcServer(parseInt(rowEl.getAttribute('index')))});if(i>0)moveUpButtonEl.addEventListener('click',()=>{
+moveUpInList(parseInt(rowEl.getAttribute('index')))})}tableNode.appendChild(rowEl)}resolve(null)});const checkForApiError=data=>new Promise((resolve,reject)=>{
+if('success'===data.status)resolve(null);else reject(new Error('PATCH API did not return success status flag'))});const setDivVisibility=data=>{
+document.getElementById('listVisibilityDiv').removeAttribute('hidden','');document.getElementById('formVisibilityDiv').setAttribute('hidden','')
 ;document.getElementById('serverPasswordWarningDiv').setAttribute('hidden','');document.getElementById('nickservCommandWarningDiv').setAttribute('hidden','');if(data.ircConnected||data.ircConnecting){
 document.getElementById('createNewButton').setAttribute('hidden','');document.getElementById('warningVisibilityDiv').removeAttribute('hidden');editable=false}else{
 document.getElementById('createNewButton').removeAttribute('hidden');document.getElementById('warningVisibilityDiv').setAttribute('hidden','');editable=true}
@@ -151,13 +160,16 @@ if(data.enableSocks5Proxy)document.getElementById('ircProxyDiv').textContent='So
 if(document.getElementById('groupInfoHiddenDiv').hasAttribute('hidden'))document.getElementById('groupInfoHiddenDiv').removeAttribute('hidden');else document.getElementById('groupInfoHiddenDiv').setAttribute('hidden','')
 });document.getElementById('replacePasswordButton').addEventListener('click',()=>{document.getElementById('passwordInputId').removeAttribute('disabled')
 ;document.getElementById('passwordInputId').value='';document.getElementById('replacePasswordButton').setAttribute('hidden','')
-;document.getElementById('serverPasswordWarningDiv').removeAttribute('hidden')});document.getElementById('replaceIdentifyCommandButton').addEventListener('click',()=>{
-document.getElementById('identifyCommandInputId').removeAttribute('disabled');document.getElementById('identifyCommandInputId').value=''
-;document.getElementById('replaceIdentifyCommandButton').setAttribute('hidden','');document.getElementById('nickservCommandWarningDiv').removeAttribute('hidden')})
-;document.getElementById('createNewButton').addEventListener('click',()=>{_clearError();fetchServerList(0,1).then(()=>fetchServerList(0,0)).then(()=>clearIrcServerForm()).then(()=>{
-document.getElementById('listVisibilityDiv').setAttribute('hidden','');document.getElementById('formVisibilityDiv').removeAttribute('hidden')
-;document.getElementById('serverPasswordWarningDiv').setAttribute('hidden','');document.getElementById('nickservCommandWarningDiv').setAttribute('hidden','')}).catch(err=>{
-_showError(err.toString()||err);console.log(err)})});document.getElementById('saveNewButton').addEventListener('click',()=>{_clearError()
+;document.getElementById('serverPasswordWarningDiv').removeAttribute('hidden')});document.getElementById('replaceSaslPasswordButton').addEventListener('click',()=>{
+document.getElementById('saslPasswordInputId').removeAttribute('disabled');document.getElementById('saslPasswordInputId').value=''
+;document.getElementById('replaceSaslPasswordButton').setAttribute('hidden','');document.getElementById('serverSaslPasswordWarningDiv').removeAttribute('hidden')})
+;document.getElementById('replaceIdentifyCommandButton').addEventListener('click',()=>{document.getElementById('identifyCommandInputId').removeAttribute('disabled')
+;document.getElementById('identifyCommandInputId').value='';document.getElementById('replaceIdentifyCommandButton').setAttribute('hidden','')
+;document.getElementById('nickservCommandWarningDiv').removeAttribute('hidden')});document.getElementById('createNewButton').addEventListener('click',()=>{_clearError()
+;fetchServerList(0,1).then(()=>fetchServerList(0,0)).then(()=>clearIrcServerForm()).then(()=>{document.getElementById('listVisibilityDiv').setAttribute('hidden','')
+;document.getElementById('formVisibilityDiv').removeAttribute('hidden');document.getElementById('serverPasswordWarningDiv').setAttribute('hidden','')
+;document.getElementById('nickservCommandWarningDiv').setAttribute('hidden','')}).catch(err=>{_showError(err.toString()||err);console.log(err)})})
+;document.getElementById('saveNewButton').addEventListener('click',()=>{_clearError()
 ;parseFormInputValues().then(data=>submitServer(data.data,'POST',-1)).then(data=>checkForApiError(data)).then(()=>fetchIrcState()).then(data=>setDivVisibility(data)).then(()=>fetchServerList(-1,-1)).then(data=>buildServerListTable(data)).catch(err=>{
 _showError(err.toString()||err);console.log(err)})});document.getElementById('saveModifiedButton').addEventListener('click',()=>{_clearError()
 ;parseFormInputValues().then(data=>submitServer(data.data,'PATCH',data.index)).then(data=>checkForApiError(data)).then(()=>fetchIrcState()).then(data=>setDivVisibility(data)).then(()=>fetchServerList(-1,-1)).then(data=>buildServerListTable(data)).catch(err=>{